optimize ui

+++ TicTacToe4x/board.cpp
@@ -56,11 +56,11 @@ bool Board::isPlayerTurn(void) {
     return b;
 }
 
-int Board::getScore(int b, int x, int y) {
+double Board::getScore(int b, int x, int y) {
     int pc = first_play_ ? OO : XX; // player color
     int cc = first_play_ ? XX : OO; // computer color
 
-    int score = -1;
+    double score = -1;
     /* TODO: use minimax-alpha-beta algorithm to optimize searching. */
     vector<Path> paths = getPaths(b, x, y);
     for (auto& path : paths) {
